Video competed
https://www.youtube.com/watch?v=JLc-hWsPTUY&list=WL&index=1&ab_channel=TylerPotts

I have completed the video but only then did I realise that it was a local app only and was missing vital components to  deploy the app.

The back image is the suitable to a full screen desktop and will need to be address.

The app on load is very plan, i have tried putting basic info in the search box place holder as a guide.

I think it would be best to get the users location and set the default to their location.